Move Nessie exception mappers to JAX-RS module
Move exception mappers created under quarkus module under the JAX-RS
`services` module, as error messages are also part of the API.
